xapp-gtk3-module.c: Use the 'realize' signal to trigger injecting the favorites:// uri into the sidebar places.
It's not important where this occurs, but doing it during the
GtkPlacesSidebar constructed vfunc now causes a deadlock in the
dconf engine between the GSettings instance for favorites, and
the one for GtkFileChooserWidget settings, created during its
construction.

I've been unable to determine any particular commit that may be
to blame, between gtk, gio, meld and dconf, and I'm unable to
reproduce this anywhere but in Meld thus far.

The 'realize' signal is sent after widget construction so is
safely outside the template machinery.

Fixes #171.